Replace obs_est code with run_symmetrized_readout and supply helper that provides a copy qc with this method.
