+++ forest/benchmarking/observable_estimation.py
@@ -22,6 +22,7 @@ from pyquil.paulis import PauliTerm, sI, is_identity
 
 from forest.benchmarking.compilation import basic_compile, _RY
 from forest.benchmarking.utils import transform_bit_moments_to_pauli
+from forest.benchmarking.pyquil_helpers import get_symmetrizing_qc
 
 if sys.version_info < (3, 7):
     from pyquil.external.dataclasses import dataclass
@@ -829,110 +830,6 @@ def generate_experiment_programs(obs_expt: ObservablesExperiment, active_reset: 
     return programs, meas_qubits
 
 
-def _flip_array_to_prog(flip_array: Tuple[bool], qubits: List[int]) -> Program:
-    """
-    Generate a pre-measurement program that flips the qubit state according to the flip_array of
-    bools.
-
-    This is used, for example, in exhaustive_symmetrization to produce programs which flip a
-    select subset of qubits immediately before measurement.
-
-    :param flip_array: tuple of booleans specifying whether the qubit in the corresponding index
-        should be flipped or not.
-    :param qubits: list specifying the qubits in order corresponding to the flip_array
-    :return: Program which flips each qubit (i.e. instructs RX(pi, q)) according to the flip_array.
-    """
-    assert len(flip_array) == len(qubits), "Mismatch of qubits and operations"
-    prog = Program()
-    for qubit, flip_output in zip(qubits, flip_array):
-        if flip_output == 0:
-            continue
-        elif flip_output == 1:
-            prog += Program(RX(pi, qubit))
-        else:
-            raise ValueError("flip_bools should only consist of 0s and/or 1s")
-    return prog
-
-
-def exhaustive_symmetrization(programs: List[Program], meas_qubits: List[List[int]]) \
-        -> Tuple[List[Program], List[List[int]], List[Tuple[bool]], List[int]]:
-    """
-    For each program in the input programs generate new programs which flip the measured qubits
-    in every possible combination in order to symmetrize readout.
-
-    The expanded list of programs is returned along with a correspondingly expanded list of
-    meas_qubits, a list of bools which indicates which qubits are flipped in each program,
-    and a list of indices which records from which program in the input programs list each
-    symmetrized program originated.
-
-    :param programs: a list of programs each of which will be symmetrized.
-    :param meas_qubits: the corresponding groups of measurement qubits for the input list of
-        programs. Only these qubits will be symmetrized over, even if the program acts on other
-        qubits.
-    :return: a list of symmetrized programs, the corresponding measurement qubits,
-        the corresponding array of bools indicating which qubits were flipped,
-        and a corresponding list of indices specifying the generating `program'
-    """
-    assert len(programs) == len(meas_qubits), 'mismatch of programs and qubits; must know which ' \
-                                              'qubits are being measured to symmetrize them.'
-
-    symm_programs = []
-    symm_meas_qs = []
-    prog_groups = []
-    flip_arrays = []
-
-    for idx, (prog, meas_qs) in enumerate(zip(programs, meas_qubits)):
-        for flip_array in itertools.product([0, 1], repeat=len(meas_qs)):
-            total_prog_symm = prog.copy()
-            prog_symm = _flip_array_to_prog(flip_array, meas_qs)
-            total_prog_symm += prog_symm
-            symm_programs.append(total_prog_symm)
-            symm_meas_qs.append(meas_qs)
-            prog_groups.append(idx)
-            flip_arrays.append(flip_array)
-
-    return symm_programs, symm_meas_qs, flip_arrays, prog_groups
-
-
-def _measure_bitstrings(qc: QuantumComputer, programs: List[Program], meas_qubits: List[List[int]],
-                        num_shots = 500, use_compiler = False) -> List[np.ndarray]:
-    """
-    Wrapper for appending measure instructions onto each program, running the program,
-    and accumulating the resulting bitarrays.
-
-    By default each program is assumed to be native quil.
-
-    :param qc: a quantum computer object on which to run each program
-    :param programs: a list of programs to run
-    :param meas_qubits: groups of qubits to measure for each program
-    :param num_shots: the number of shots to run for each program
-    :return: a len(programs) long list of num_shots by num_meas_qubits bit arrays of results for
-        each program.
-    """
-    assert len(programs) == len(meas_qubits), 'The qubits to measure must be specified for each ' \
-                                              'program, one list of qubits per program.'
-
-    results = []
-    for program, qubits in zip(programs, meas_qubits):
-        if len(qubits) == 0:
-            # corresponds to measuring identity; no program needs to be run.
-            results.append(np.array([[]]))
-            continue
-        # copy the program so the original is not mutated
-        prog = program.copy()
-        ro = prog.declare('ro', 'BIT', len(qubits))
-        for idx, q in enumerate(qubits):
-            prog += MEASURE(q, ro[idx])
-
-        prog.wrap_in_numshots_loop(num_shots)
-        if use_compiler:
-            prog = qc.quil_to_native_quil(prog)
-        exe = qc.compiler.native_quil_to_executable(prog)
-        shots = qc.run(exe)
-        results.append(shots)
-    return results
-
-
 def shots_to_obs_moments(bitarray: np.ndarray, qubits: List[int], observable: PauliTerm,
                          use_beta_dist_unbiased_prior: bool = False) -> Tuple[float, float]:
     """
@@ -985,39 +882,8 @@ def shots_to_obs_moments(bitarray: np.ndarray, qubits: List[int], observable: Pa
     return obs_mean, obs_var
 
 
-def consolidate_symmetrization_outputs(outputs: List[np.ndarray], flip_arrays: List[Tuple[bool]],
-                                       groups: List[int]) -> List[np.ndarray]:
-    """
-    Given bitarray results from a series of symmetrization programs, appropriately flip output
-    bits and consolidate results into new bitarrays.
-
-    :param outputs: a list of the raw bitarrays resulting from running a list of symmetrized
-        programs; for example, the results returned from _measure_bitstrings
-    :param flip_arrays: a list of boolean arrays in one-to-one correspondence with the list of
-        outputs indicating which qubits where flipped before each bitarray was measured.
-    :param groups: the group from which each symmetrized program was generated. E.g. if only one
-        program was symmetrized then groups is simply [0] * len(outputs). The length of the
-        returned consolidated outputs is exactly the number of distinct integers in groups.
-    :return: a list of the consolidated bitarray outputs which can be treated as the symmetrized
-        outputs of the original programs passed into a symmetrization method. See
-        estimate_observables for example usage.
-    """
-    assert len(outputs) == len(groups) == len(flip_arrays)
-
-    output = {group: [] for group in set(groups)}
-    for bitarray, group, flip_array in zip(outputs, groups, flip_arrays):
-        if len(flip_array) == 0:
-            # happens when measuring identity.
-            # TODO: better way of handling identity measurement? (in _measure_bitstrings too)
-            output[group].append(bitarray)
-        else:
-            output[group].append(bitarray ^ flip_array)
-
-    return [np.vstack(output[group]) for group in sorted(list(set(groups)))]
-
-
 def estimate_observables(qc: QuantumComputer, obs_expt: ObservablesExperiment,
-                         num_shots: int = 500, symmetrization_method: Callable = None,
+                         num_shots: int = 500, symm_type: int = 0, use_basic_compile = True,
                          active_reset: bool = False) -> Iterable[ExperimentResult]:
     """
     Standard wrapper for estimating the observables in an ObservableExperiment.
@@ -1038,24 +904,20 @@ def estimate_observables(qc: QuantumComputer, obs_expt: ObservablesExperiment,
         observable of obs_expt.
     :param obs_expt: a single ObservablesExperiment with settings pre-grouped as desired.
     :param num_shots: the number of shots to run each program or each symmetrized program.
-    :param symmetrization_method: if not None, this can be used to symmetrize each program to
-        remove unwanted classical correlations in the qubit measurement readout.
+    :param symm_type: the type of symmetrization
+    :param use_basic_compile: by default use basic compile instead of
+        qc.compiler.quil_to_native_quil which potentially optimizes away gates.
     :param active_reset: whether or not to begin the program by actively resetting. If true,
         execution of each of the returned programs in a loop on the QPU will generally be faster.
     :return: all of the ExperimentResults which hold an estimate of each observable of obs_expt
     """
     programs, meas_qubits = generate_experiment_programs(obs_expt, active_reset)
 
-    if symmetrization_method is not None:
-        programs, symm_qubits, flip_arrays, prog_groups = symmetrization_method(programs,
-                                                                                meas_qubits)
-        symm_outputs = _measure_bitstrings(qc, programs, symm_qubits, num_shots)
-        results = consolidate_symmetrization_outputs(symm_outputs, flip_arrays, prog_groups)
-    else:
-        results = _measure_bitstrings(qc, programs, meas_qubits, num_shots)
+    # TODO: remove hack to use symmetrization methods and skip compilation
+    symm_qc = get_symmetrizing_qc(qc, use_basic_compile)
 
-    assert len(results) == len(meas_qubits) == len(obs_expt)
-    for bitarray, meas_qs, settings in zip(results, meas_qubits, obs_expt):
+    for program, meas_qs, settings in zip(programs, meas_qubits, obs_expt):
+        bitarray = symm_qc.run_symmetrized_readout(program, num_shots, symm_type, meas_qs)
 
         for setting in settings:
             observable = setting.observable
@@ -1071,15 +933,21 @@ def estimate_observables(qc: QuantumComputer, obs_expt: ObservablesExperiment,
             )
 
 
-def get_calibration_program(observable: PauliTerm, noisy_program: Program = None) -> Program:
+def get_calibration_program(observable: PauliTerm, noisy_program: Program = None,
+                            active_reset: bool = False) -> Program:
     """
     Program required for calibrating the given observable.
 
     :param observable: observable to calibrate
     :param noisy_program: a program with readout and gate noise defined; only useful for QVM
+    :param active_reset: whether or not to begin the program by actively resetting. If true,
+        execution of each of the returned programs in a loop on the QPU will generally be faster.
     :return: Program performing the calibration
     """
     calibr_prog = Program()
+
+    if active_reset:
+        calibr_prog += RESET()
 
     # Inherit any noisy attributes from noisy_program, including gate definitions
     # and applications which can be handy in simulating noisy channels
@@ -1103,9 +971,9 @@ def get_calibration_program(observable: PauliTerm, noisy_program: Program = None
 
 
 def calibrate_observable_estimates(qc: QuantumComputer, expt_results: List[ExperimentResult],
-                                   symmetrization_method: Callable = exhaustive_symmetrization,
-                                   num_shots: int = 500, noisy_program: Program = None) \
-        -> Iterable[ExperimentResult]:
+                                   symm_type: int = -1, num_shots: int = 500,
+                                   noisy_program: Program = None, use_basic_compile = True,
+                                   active_reset: bool = False) -> Iterable[ExperimentResult]:
     """
     Calibrates the expectation and std_err of the input expt_results and updates those estimates.
 
@@ -1118,10 +986,15 @@ def calibrate_observable_estimates(qc: QuantumComputer, expt_results: List[Exper
     :param qc: a quantum computer object on which to run the programs necessary to calibrate each
         result.
     :param expt_results: a list of results, each of which will be separately calibrated.
-    :param symmetrization_method: by default every eigenvector of each observable is measured.
+    :param symm_type: the type of symmetrization; by default every eigenvector of each observable
+        is measured.
     :param num_shots: the number of shots to run for each eigenvector
     :param noisy_program: an optional program from which to inherit a noise model; only relevant
         for running on a QVM
+    :param use_basic_compile: by default use basic compile instead of
+        qc.compiler.quil_to_native_quil which potentially optimizes away gates.
+    :param active_reset: whether or not to begin the program by actively resetting. If true,
+        execution of each of the returned programs in a loop on the QPU will generally be faster.
     :return: a copy of the input results with updated estimates and calibration results.
     """
     observables = [copy(res.setting.observable) for res in expt_results]
@@ -1129,16 +1002,16 @@ def calibrate_observable_estimates(qc: QuantumComputer, expt_results: List[Exper
         obs.coefficient = 1.
     observables = list(set(observables))  # get unique observables that will need to be calibrated
 
-    programs = [get_calibration_program(obs, noisy_program) for obs in observables]
+    programs = [get_calibration_program(obs, noisy_program, active_reset) for obs in observables]
     meas_qubits = [obs.get_qubits() for obs in observables]
 
-    symm_progs, symm_qubits, flip_array, prog_groups = symmetrization_method(programs, meas_qubits)
-    symm_outputs = _measure_bitstrings(qc, symm_progs, symm_qubits, num_shots)
-    results = consolidate_symmetrization_outputs(symm_outputs, flip_array, prog_groups)
+    # TODO: remove hack to use symmetrization methods and skip compilation
+    symm_qc = get_symmetrizing_qc(qc, use_basic_compile)
 
-    assert len(results) == len(meas_qubits) == len(observables)
     calibrations = {}
-    for bitarray, meas_qs, obs in zip(results, meas_qubits, observables):
+    for program, meas_qs, obs in zip(programs, meas_qubits, observables):
+        bitarray = symm_qc.run_symmetrized_readout(program, num_shots, symm_type, meas_qs)
+
         # Obtain statistics from result of experiment
         obs_mean, obs_var = shots_to_obs_moments(bitarray, meas_qs, obs)
         calibrations[obs.operations_as_set()] = (obs_mean, obs_var, len(bitarray))
