+++ forest/benchmarking/pyquil_helpers/__init__.py
@@ -1,9 +1,1 @@
-from ._quantum_computer import QuantumComputer, get_qc, list_quantum_computers
-
-from ._symmetrization_helpers import (_symmetrization,
-                                      _flip_array_to_prog,
-                                      _construct_orthogonal_array,
-                                      _construct_strength_two_orthogonal_array,
-                                      _construct_strength_three_orthogonal_array,
-                                      _measure_bitstrings, _consolidate_symmetrization_outputs,
-                                      _check_min_num_trials_for_symmetrized_readout)
+from ._symmetrization_helpers import get_symmetrizing_qc
